fixup! first increment
