--- cmd/shortener/README.md
@@ -1,13 +1,0 @@
-# cmd/shortener
-
-В данной директории содержится код, который скомпилируется в бинарное приложение.
-
-Рекомендуется помещать только код, необходимый для запуска приложения, но не бизнес-логику.
-
-Название директории должно соответствовать названию приложения.
-
-Директория `cmd/shortener` содержит:
-- точку входа в приложение (функция `main`)
-- инициализацию зависимостей (можно вынести в отдельный пакет `internal/app`)
-- настройку и запуск HTTP-сервера (можно вынести в отдельный пакет `internal/router`)
-- обработку сигналов завершения работы приложения
