+++ go.mod
@@ -1,3 +1,11 @@
-module github.com/kujoki/go-musthave-service.git
+module github.com/kujoki/go-musthave-service
 
 go 1.23.10
+
+require github.com/stretchr/testify v1.10.0
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
